Install OfficeToPDF

Downloads the executable file OfficeToPDF.exe from https://github.com/cognidox/OfficeToPDF/releases.

install_otp(dir, url, method, quiet = TRUE, ...)

Arguments

dir

Directory into which to download the OfficeToPDF.exe. If dir is missing then system.file(package = "accessr") is used. This should enable OfficeToPDF.exe to be found but there may be a problem if .libPaths refers to a cloud storage directory, such as OneDrive.

url

URL from which to download OfficeToPDF.exe. If this is missing then https://github.com/cognidox/OfficeToPDF/releases/download/v1.9.0.2/OfficeToPDF.exe is used.

method

Passed to download.file. If missing then getOption("download.file.method") is used to set it.

quiet

Passed to download.file. quiet controls whether messages are printed during the download.

...

Additional arguments to be passed to download.file.

Value

See the Value section of download.file.

Details

The argument mode = "wb" to download.file is hard-coded, so that binary transfer is forced.

Note

OfficeToPDF.exe is used to create PDFs files from Word documents in rmd2word.

Examples

install_otp()