Enable ioslides slide_level = 1

Enables the argument slide_level = 1 to ioslides_presentation to set the slide separator as the level one header #.

ioslides_level(one = TRUE)

Arguments

one

A logical scalar. If one = TRUE a new Lua filter is used that enables slide_level = 1 to set the slide separator as the level one header #. A modified default.css css file is also used that adjusts the font sizes for the header levels accordingly, so that the level one header has a larger font than the level two header. If one = FALSE then the original rmarkdown versions of the Lua filter and the css file are reinstated.

Value

A logical scalar returned from file.copy.

Details

The function ioslides_presentation has an argument slide_level that sets the header level used as a slide separator. The Lua filter ioslides_presentation.lua in the rmarkdown package uses any content between headers of level slide_level to create a segue slide, which has a grey background and is intended only to contain a section heading.

In particular, under the default, slide_level = 2, content between a level one header # and the next level two header ## is formatted as a separate grey segue slide. If we do not want segue slides then we must avoid using level one headers. However, for reasons of document accessibility, we may want to use level one headers to separate slides. For example, if we wish to create an ioslides presentation and a Word document from the same source Rmd file then the Word document will only meet fully accessibility requirements if the headings in the document start at level one.

In rmarkdown version 2.25, passing slide_level= 1 to ioslides_presentation does not force a new non-segue slide when a level one header # is used: it places all content between # and the next ## on a grey segue slide and the behaviour content of the resulting slides is not desirable. Executing the ioslides_level function replaces rmarkdown's Lua filter ioslides_presentation.lua with one that has been modified so that passing slide_level = 1 will have this effect. For values of slide_level greater than or equal to 2 ioslides_presentation will behave as usual, except that the default font sizes will be altered owing to the new default.css file. Therefore, it may be better to use ioslides(one = TRUE) with an Rmd file in which level one headers are used to separate slides.

Examples

ioslides_level()