Converts rmarkdown code to Word and PDF documents

Creates accessible PDF documents by creating Word documents using rmarkdown code and then producing PDF files.

rmd2pdf(x, doc = NULL, dir = NULL, ...)

pns(x, ...)

pnd(x, ...)

Arguments

x
A character vector containing the names of the .Rmd files to convert (no extension) if they are in the current working directory, or paths to the files, either absolute or relative to the current working directory, e.g., DIRECTORY/file1. The .docx and .pdf files are created in the same directory as their respective .Rmd file.

doc
An optional argument to identify a template Word document, or respective documents for the files in x, from which the fonts, margins etc in the output Word documents will based. If doc is not supplied then a default is used. Different templates may be used for different files. rep_len(doc, length(x)) is used to force length(doc) to have the same length as x. A component equal to "default" may be used to use the default Word template. See Details for information about how to identify the locations of the Word template files.

dir
A path to the directory in which the file officetopdf.exe sits. This is not needed if this file sits in the current working directory or a directory in the list returned by searchpaths(). Otherwise, it may be a path relative to the current working directory or an absolute path.

...
Additional arguments to be passed to system. The argument wait determines whether or not R will wait for the PDF files to be produced. In the default wait = TRUE case a warning will be given if any of the PDF files could not be produced. This will occur if there is an existing PDF file of the same name open in another application. If wait = FALSE then no warnings will be produced.

Value

A vector containing the values returned from system is returned invisibly. Note that if wait = FALSE then these values will be 0 (the success value) even if some of the PDF files could not be produced. The error code 17234 indicates that a PDF file was open in another application.

Details

The simplest setup is to have the .Rmd files and the Word template (if used) and officetopdf.exe in the current working directory.

It is possible to have the .Rmd files in different directories, but if any non-"default" values in doc must be such that the reference_docx argument of word_document finds a template Word file. If the template is in the same directory as its respective .Rmd component in x then the filename, e.g. template.docx will suffice. Otherwise, a path to the template should be given, either relative to the directory in which the .Rmd file sits, or an absolute path.

The render function creates a Word file from each input .Rmd file. Then OfficeToPDF is used to convert the Word file to a PDF file. The file officetopdf.exe needs to be downloaded from the OfficeToPDF releases page and placed in the directory specified by the argument dir, or in a directory that is in the list returned by searchpaths. If officetopdf.exe cannot be found then an error is thrown.

pns and pnd are convenience functions for the package maintainer.

Examples

if (FALSE) {
# All files in the current working directory
rmd2pdf(c("file1", "file2"), doc = "template.docx")
}