Converts R markdown code to Word and PDF documents

Creates accessible PDF documents by creating Word documents using R markdown code and then producing PDF files. Zip archives of the PDF files may be created.

rmd2word(
  x,
  doc = "accessr",
  dir,
  zip = TRUE,
  add = FALSE,
  quiet = TRUE,
  rm_word = FALSE,
  rm_pdf = FALSE,
  inc_word = FALSE,
  ...
)

Arguments

x
A character vector containing the names (no extension) of the .Rmd files to convert if they are in the current working directory, or paths to the files, either absolute or relative to the current working directory, e.g., DIRECTORY/file1. The .docx and .pdf files are created in the same directory as their respective .Rmd file. If x is missing then a PDF file is created from each of the .Rmd files in the current working directory.

doc
An optional character vector (including the file extension) to specify template Word documents on which to base the style of the respective output Word documents. This determines what is passed as the argument reference_docx to word_document, via rdocx_document. Different templates may be used for different files. rep_len(doc, length(x)) is used to force length(doc) to have the same length as x. A component equal to "officedown" specifies rdocx_document's default. A component equal to "accessr" specifies accessr's internal template file, which has narrower margins and darker blue fonts for titles and hyperlinks, to avoid contrast issues. To use your own template(s), provide their filenames. See Details for more information.

dir
A path to the directory in which the file OfficeToPDF.exe sits. This is not needed if this file sits in the current working directory or a directory in the list returned by searchpaths(). Otherwise, it may be a path relative to the current working directory or an absolute path.

zip
A logical scalar or character vector indicating whether PDF files should be put into a zip archive. If zip = FALSE then no zip archive is created. Otherwise, an archive is created in each unique directory involved in x. If zip = TRUE then any archive created has the name accessr_word.zip. If zip is a character vector of zip file names (no extension) then these names are used to name the zip archives. The names are recycled to the length of the number of unique directories if necessary.

add
A logical scalar that determines what happens if the output zip file already exists. If add = TRUE then files are added to the zip file and if add = FALSE then the zip file is deleted and will only contain newly-created files.

rm_word
A logical scalar. If rm_word = TRUE then all the Word files created are deleted. Otherwise, they are not deleted.

rm_pdf
A logical scalar. If rm_pdf = TRUE and a zip archive of PDF files is produced then the individual PDF files are deleted. Otherwise, they are not deleted.

inc_word
A logical scalar. If inc_word = TRUE then the Word files are included in the zip file created. Otherwise, they are not included.

...
Additional arguments to be passed to word_document.

Value

In addition to creating the Word and PDF files, and perhaps zip files, a list containing the following vector components is returned invisibly:

error_codes

numeric values returned from system. If wait = FALSE then these values will be 0 (the success value) even if some of the PDF files could not be produced. The error code 17234 indicates that a PDF file was open in another application.

files

(absolute) paths and file names of the files added to a zip file.

zips

(relative) paths and names of all the zip files.

Details

Information such as title, author, lang etc in the YAML header in the Rmd file are used but output is ignored.

The simplest setup is to have the .Rmd files and the Word template (if used) and OfficeToPDF.exe in the current working directory.

It is possible to have the .Rmd files in different directories, but any non-"default" values in doc must be such that the reference_docx argument of word_document finds a template Word file. If the template is in the same directory as its respective .Rmd component in x then the filename, e.g. "template.docx" will suffice. Otherwise, a path to the template should be given, either relative to the directory in which the .Rmd file sits, or an absolute path.

For information on how to create a template Word document Happy collaboration with Rmd to docx.

The render function creates a Word file from each input .Rmd file. Then OfficeToPDF is used to convert the Word file to a PDF file. The file OfficeToPDF.exe needs to be downloaded from the OfficeToPDF releases page and placed in the directory specified by the argument dir, or in a directory that is in the list returned by searchpaths. If OfficeToPDF.exe cannot be found then an error is thrown. A warning will be given if any of the PDF files could not be produced. This will occur if there is an existing PDF file of the same name open in another application.

References

Layton, Richard. (2015) Happy collaboration with Rmd to docx. R Markdown from RStudio article. https://rmarkdown.rstudio.com/articles_docx.html

Examples

if (FALSE) {
# All files in the current working directory
rmd2word(c("TEST/file1", "TEST/file2"), doc = "template.docx")
}