accessr: Command Line Tools to Produce Accessible Documents using R markdown

This package provides functions to produce accessible html and PDF documents from input R markdown files. Currently, accessr only provides the option to produce accessible PDF files on a Windows Operating System. One aspect of accessibility is providing headings that are recognised by a screen reader, providing a navigational tool for a blind or partially-sighted person. A main aim is to enable documents of different formats to be produced from a single R markdown source file using a single function call. A user-supplied template Word document can be used to determine the formatting of the output Word document. Similar functions produce html slides and html documents. A zip file containing multiple files can be produced. The option to print html output to (non-accessible) PDF files is also available.

Details

See the accessr package page on Github for more information. An example Rmd file is available at system.file(package = "accessr", "examples", "example.Rmd").

On a Windows Operating System, Accessible PDF documents are produced by creating Word documents from R markdown files and then PDF documents from these Word documents. The first step uses the render function from the rmarkdown package and the rdocx_document function from the officedown package. The second step uses OfficeToPDF.

The main functions in accessr are:

rmd2many: create HTML slides, PDF slides, Word and PDF documents from a single R markdown file.

rmd2word: create Word documents and accessible PDF files.

install_otp: convenience function to install OfficeToPDF, to create PDF files from Word documents in rmd2word. ext_img: a function to enable the knitr chunk options out.width and/or out.height to work when the output format is a Word document

rmd2ioslides: create ioslides presentations and perhaps print to (non-accessible) PDF documents.

ioslides_level: enable the use of the level one header # to separate slides in an ioslides presentation without producing segue slides.

rmd2slidy: create slidy presentations and perhaps print to (non-accessible) PDF documents..

rmd2html: create html documents and perhaps print to (non-accessible) PDF documents.

The rmd2? functions provide the option to create a zip archive containing the output files. All the .Rmd files in a directory can be processed with one function call. Information such as title, author, lang etc in the YAML header in the Rmd file are used but output is ignored.

References

David Gohel and Noam Ross (2021). officedown: Enhanced 'R Markdown' Format for 'Word' and 'PowerPoint'. R package version 0.2.2. https://CRAN.R-project.org/package=officedown

JJ Allaire, Yihui Xie, Jonathan McPherson, Javier Luraschi, Kevin Ushey, Aron Atkins, Hadley Wickham, Joe Cheng, Winston Chang and Richard Iannone (2021). rmarkdown: Dynamic Documents for R. R package version 2.9. https://rmarkdown.rstudio.com.

See also

Author

Maintainer: Paul J. Northrop p.northrop@ucl.ac.uk [copyright holder]